Implemented variants - BackslashLogicals
se.txt
BackslashLogicals: 0
Use & or | for AND, OR
db.txt
<test-><ifcb:a|b|c>1
<test-><app:*db.txt - Notepad | db.txt - Notepad> 1

BackslashLogicals: 1
Use \& or \| for AND, OR
<test-><ifcb:a\|b\|c>1

<<:> Stream options
Use \# for color
<test-><<: \1 test>

Use \c for clipboard
<test-><<:\c>

Use \' for a blank output
<test-><<:\'test>

Implemented <xy:><rgb:> two-in-one
0 0 is for 🖱️ X Y
e.g.
<test-><rgb: 255 255 255 0 0>1
<test-><RGB: 255 255 255 0 0>1

Extended else connect
Use " " for <true:> or <false:> connect option:
<test-><ifcb: a|b|c, 1, 1, true: false:>
<true:>1
<false:>0
Previous:
<test-><ifcb: a, 1, 1, false:>1

Use "<" for reconnect option
<test-><ifcb: a | b | c, 3, 1000, <true: <false:>2

Same signature in:
<app:>
<App:>
<rgb:>
<RGB:>
<ifcb:>
<ifs:>

Print values
<app>
<rgb>
<xy>
e.g.
<test-><lt:>rgb:<rgb> <xy>>
<lt:><shift>,<shift->